+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+#### 2.1.0-rc.1
+
+* Updated build process to webpack 4
+* Updated build process to Babel 7
+* Updated build process to use uglify webpack plugin over cli
+* Updated unit test to Jest 24
+
 #### 1.1.0
 
 * Support UMD